+++ packages/browser-wallet/src/popup/pages/Account/Tokens/TokenBalance.tsx
@@ -1,16 +1,19 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { Atom, useAtomValue } from 'jotai';
 import { addThousandSeparators, toFraction } from 'wallet-common-helpers';
+import { ContractBalances } from '@shared/utils/token-helpers';
 
 type BalanceProps = {
-    atom: Atom<Promise<bigint>>;
+    atom: Atom<Promise<ContractBalances>>;
+    id: string;
     decimals: number;
+    children?(balance: bigint): ReactNode;
 };
 
-export default function TokenBalance({ atom, decimals }: BalanceProps) {
-    const balance = useAtomValue(atom);
+export default function TokenBalance({ atom, decimals, children, id }: BalanceProps) {
+    const balances = useAtomValue(atom);
     const getFraction = toFraction(10 ** decimals);
-    const renderBalance = (value: bigint) => addThousandSeparators(getFraction(value));
+    const renderBalance = children ?? ((value: bigint) => addThousandSeparators(getFraction(value)));
 
-    return <>{renderBalance(balance)}</>;
+    return <>{renderBalance(balances[id])}</>;
 }