+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/TransactionList.tsx
@@ -13,6 +13,8 @@ import { useAtomValue, useSetAtom } from 'jotai';
 import Button from '@popup/shared/Button';
 import { networkConfigurationAtom } from '@popup/store/settings';
 import { isMainnet } from '@shared/utils/network-helpers';
+import { addPendingTransactionAtom, selectedPendingTransactionsAtom } from '@popup/store/transactions';
+import { useUpdateAtom } from 'jotai/utils';
 import useTransactionGroups, { TransactionsByDateTuple } from './useTransactionGroups';
 import TransactionElement, { transactionElementHeight } from './TransactionElement';
 
@@ -185,7 +187,18 @@ export default function TransactionList({ onTransactionClick }: TransactionListP
     const { t } = useTranslation('transactionLog');
     const account = useSelectedCredential();
     const accountAddress = useMemo(() => account?.address, [account]);
+    const pendingTransactions = useAtomValue(selectedPendingTransactionsAtom);
+    const addPendingTransaction = useUpdateAtom(addPendingTransactionAtom);
     const [transactions, setTransactions] = useState<BrowserWalletTransaction[]>([]);
+    const allTransactions = useMemo(
+        () => [
+            ...pendingTransactions.filter(
+                (ta) => !transactions.some((tb) => ta.transactionHash === tb.transactionHash)
+            ),
+            ...transactions,
+        ],
+        [transactions, pendingTransactions]
+    );
     const [isNextPageLoading, setIsNextPageLoading] = useState<boolean>(false);
     const [hasNextPage, setHasNextPage] = useState<boolean>(true);
     const addToast = useSetAtom(addToastAtom);
@@ -226,9 +239,11 @@ export default function TransactionList({ onTransactionClick }: TransactionListP
         getTransactions(address, transactionResultLimit, 'descending', fromId)
             .then((transactionResult) => {
                 setHasNextPage(transactionResult.full);
+
                 const updatedTransactions = appendTransactions
                     ? transactions.concat(transactionResult.transactions)
                     : transactionResult.transactions;
+
                 setTransactions(updatedTransactions);
                 setIsNextPageLoading(false);
             })
@@ -260,14 +275,12 @@ export default function TransactionList({ onTransactionClick }: TransactionListP
         setAmount(accountInfo?.accountAmount);
     }, [accountInfo?.accountAmount]);
 
-    async function ccdDrop(address: string) {
-        getCcdDrop(address).then((dropTransaction) => {
-            setTransactions([dropTransaction]);
-        });
+    function ccdDrop(address: string) {
+        getCcdDrop(address).then(addPendingTransaction);
     }
 
     let transactionListComponent;
-    if (transactions.length === 0) {
+    if (allTransactions.length === 0) {
         if (isNextPageLoading) {
             transactionListComponent = null;
         } else {
@@ -291,16 +304,15 @@ export default function TransactionList({ onTransactionClick }: TransactionListP
             );
         }
     } else {
-        // If the first transaction in the list changes, then we need to reload the entire
-        // list, otherwise the infinite list component bugs out displaying the transactions
-        // correctly.
-        const listKey = transactions[0].transactionHash ?? transactions[0].id;
+        const txKey = transactions[0]?.transactionHash || transactions[0]?.id || '';
+        const listKey = `${accountAddress}${txKey}`;
+
         transactionListComponent = (
             <div className="transaction-list__scroll">
                 <InfiniteTransactionList
                     key={listKey}
                     accountAddress={accountAddress}
-                    transactions={transactions}
+                    transactions={allTransactions}
                     loadNextPage={loadNextPage}
                     hasNextPage={hasNextPage}
                     isNextPageLoading={isNextPageLoading}