+++ packages/browser-wallet/src/popup/pages/Account/i18n/da.ts
@@ -64,6 +64,8 @@ const t: typeof en = {
         addToken: 'Tilføj token',
         duplicateId: 'Token er allerede i listen',
         updateTokens: 'Opdater tokens',
+        unownedUnique: 'ejer ikke',
+        hexId: 'Id skal være HEX encodet',
     },
     accountPending: 'Denne konto er stadig ved at blive oprettet.',
     request: 'Opret konto',