+++ packages/browser-wallet/src/shared/utils/function-helpers.ts
@@ -1,3 +1,5 @@
+import { sleep } from 'wallet-common-helpers';
+
 /**
  * Takes a function which might return undefined, and throws an error instead.
  */
@@ -8,7 +10,7 @@ export function throwIfUndefined<T extends any[], V>(
 ): (...inputs: T) => V {
     return (...inputs) => {
         const result = func(...inputs);
-        if (!result) {
+        if (result === undefined) {
             throw new Error(getErrorMessage(...inputs));
         }
         return result;
@@ -19,3 +21,17 @@ export function throwIfUndefined<T extends any[], V>(
 export function not<T extends any[]>(predicate: (...inputs: T) => boolean) {
     return (...inputs: T): boolean => !predicate(...inputs);
 }
+
+type ShouldLoop = boolean;
+
+/**
+ * Takes a loopFun, that runs each interval, for as long as the loopFun resolves to true
+ */
+export const loop = async (intervalMS: number, loopFun: () => Promise<ShouldLoop>) => {
+    const run = async () => {
+        if (await loopFun()) {
+            await sleep(intervalMS).then(run);
+        }
+    };
+    await run();
+};