+++ packages/browser-wallet/src/shared/utils/token-helpers.ts
@@ -3,12 +3,6 @@ import { AccountAddress, InstanceInfo, JsonRpcClient } from '@concordium/web-sdk
 import uleb128 from 'leb128/unsigned';
 import { TokenMetadata } from '@shared/storage/types';
 
-export interface TokenIdAndMetadata {
-    id: string;
-    metadataLink: string;
-    metadata: TokenMetadata;
-}
-
 export interface ContractDetails {
     contractName: string;
     index: bigint;
@@ -38,6 +32,20 @@ export function getMetadataParameter(id: string): Buffer {
 }
 
 /**
+ * Returns the url for the token metadata.
+ * returnValue is assumed to be a HEX-encoded string.
+ */
+function deserializeTokenMetadataReturnValue(returnValue: string) {
+    const bufferStream = Buffer.from(returnValue, 'hex');
+    const length = bufferStream.readUInt16LE(2);
+    const url = bufferStream.slice(4, 4 + length).toString('utf8');
+    return url;
+}
+
+// Methods in the CIS2 standard.
+const CIS2Methods = ['balanceOf', 'transfer', 'tokenMetadata', 'operatorOf', 'updateOperator'];
+
+/**
  * Confirms that the given smart contract instance is CIS-2 compliant
  */
 export async function confirmCIS2Contract(
@@ -45,21 +53,16 @@ export async function confirmCIS2Contract(
     instanceInfo: InstanceInfo,
     { contractName, index, subindex }: ContractDetails
 ): Promise<string | undefined> {
-    if (!instanceInfo.methods.includes(`${contractName}.supports`)) {
+    const cis0SupportsMethod = `${contractName}.supports`;
+    if (!instanceInfo.methods.includes(cis0SupportsMethod)) {
         return 'Chosen contract does not support CIS-0';
     }
-    if (
-        !(
-            instanceInfo.methods.includes(`${contractName}.balanceOf`) &&
-            instanceInfo.methods.includes(`${contractName}.transfer`) &&
-            instanceInfo.methods.includes(`${contractName}.tokenMetadata`)
-        )
-    ) {
+    if (!CIS2Methods.every((method) => instanceInfo.methods.includes(`${contractName}.${method}`))) {
         return 'Chosen contract does not expose required endpoints';
     }
     const supports = await client.invokeContract({
         contract: { index, subindex },
-        method: `${contractName}.supports`,
+        method: cis0SupportsMethod,
         parameter: getCIS2Identifier(),
     });
     if (!supports || supports.tag === 'failure') {
@@ -79,7 +82,7 @@ export function getTokenUrl(
     id: string,
     { contractName, index, subindex }: ContractDetails
 ): Promise<string> {
-    return new Promise((resolve) => {
+    return new Promise((resolve, reject) => {
         client
             .invokeContract({
                 contract: { index, subindex },
@@ -88,12 +91,10 @@ export function getTokenUrl(
             })
             .then((returnValue) => {
                 if (returnValue && returnValue.tag === 'success' && returnValue.returnValue) {
-                    const bufferStream = Buffer.from(returnValue.returnValue, 'hex');
-                    const length = bufferStream.readUInt16LE(2);
-                    const url = bufferStream.slice(4, 4 + length).toString('utf8');
-                    resolve(url);
+                    resolve(deserializeTokenMetadataReturnValue(returnValue.returnValue));
                 } else {
-                    // Throw an error;
+                    // TODO: perhaps we need to make this error more precise
+                    reject(new Error('Token does not exist in this contract'));
                 }
             });
     });
@@ -126,53 +127,83 @@ export async function getTokenMetadata(tokenUrl: string): Promise<TokenMetadata>
 /**
  * Serialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#id3
  */
-const serializeBalanceParameter = (tokenIndex: string, accountAddress: string) => {
+const serializeBalanceParameter = (tokenIds: string[], accountAddress: string) => {
     const queries = Buffer.alloc(2);
-    queries.writeUInt16LE(1, 0); // 1 query
+    queries.writeUInt16LE(tokenIds.length, 0);
 
-    const token = Buffer.from(tokenIndex, 'hex');
-    const tokenLength = Buffer.alloc(1);
-    tokenLength.writeUInt8(token.length, 0);
+    const tokens = tokenIds.reduce((acc, t) => {
+        const token = Buffer.from(t, 'hex');
+        const tokenLength = Buffer.alloc(1);
+        tokenLength.writeUInt8(token.length, 0);
 
-    const addressType = Buffer.alloc(1); // Account address type
-    const address = new AccountAddress(accountAddress).decodedAddress;
+        const addressType = Buffer.alloc(1); // Account address type
+        const address = new AccountAddress(accountAddress).decodedAddress;
 
-    return Buffer.concat([queries, tokenLength, token, addressType, address]);
+        return Buffer.concat([acc, tokenLength, token, addressType, address]);
+    }, queries);
+
+    return tokens;
 };
 
 /**
  * Deserialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#response
  */
-const deserializeBalanceAmount = (value: string): bigint => {
+const deserializeBalanceAmounts = (value: string): bigint[] => {
     const buf = Buffer.from(value, 'hex');
-    const amount = uleb128.decode(buf.slice(2)); // ignore first 2 bytes as we only expect 1 tokenamount
+    const n = buf.readUInt16LE(0);
+    let cursor = 2; // First 2 bytes hold number of token amounts included in response.
+    const amounts: bigint[] = [];
 
-    return BigInt(amount);
+    // eslint-disable-next-line no-plusplus
+    for (let i = 0; i < n; i++) {
+        const end = buf.slice(cursor).findIndex((b) => b < 2 ** 7) + 1; // Find the first byte with most significant bit not set, signaling the last byte in the leb128 slice.
+
+        const amount = uleb128.decode(buf.slice(cursor, cursor + end));
+        amounts.push(BigInt(amount));
+
+        cursor += end;
+    }
+
+    return amounts;
 };
 
-export const getTokenBalance = async (
+export type ContractBalances = Record<string, bigint>;
+
+export const getContractBalances = async (
     client: JsonRpcClient,
     contractIndex: string,
-    tokenIndex: string,
+    tokenIds: string[],
     accountAddress: string
-): Promise<bigint> => {
+): Promise<ContractBalances> => {
     const index = BigInt(contractIndex);
     const subindex = 0n;
     const instanceInfo = await client.getInstanceInfo({ index, subindex });
 
     if (instanceInfo === undefined) {
-        return 0n;
+        return {};
     }
 
     const result = await client.invokeContract({
         contract: { index, subindex },
         method: `${instanceInfo.name.substring(5)}.balanceOf`,
-        parameter: serializeBalanceParameter(tokenIndex, accountAddress),
+        parameter: serializeBalanceParameter(tokenIds, accountAddress),
     });
 
     if (result === undefined || result.tag === 'failure' || result.returnValue === undefined) {
-        return 0n;
+        return {};
     }
 
-    return deserializeBalanceAmount(result.returnValue);
+    const amounts = deserializeBalanceAmounts(result.returnValue);
+
+    if (amounts.length !== tokenIds.length) {
+        throw new Error('Mismatch between length of requested tokens and token amounts in response.');
+    }
+
+    return tokenIds.reduce<ContractBalances>(
+        (acc, cur, i) => ({
+            ...acc,
+            [cur]: amounts[i],
+        }),
+        {}
+    );
 };