+++ packages/browser-wallet/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   A toast is now shown when the passcode is successfully updated.
+-   Shows submitted transactions in the transaction list.
 
 ### Changed